# Conflicts:
#	common/src/main/java/net/mca/entity/ai/brain/VillagerTasksMCA.java
#	common/src/main/java/net/mca/entity/ai/brain/tasks/WanderOrTeleportToTargetTask.java
#	forge/src/main/resources/META-INF/mods.toml

+++ common/src/main/java/net/mca/item/RelationshipItem.java
@@ -11,7 +11,7 @@ public abstract class RelationshipItem extends TooltippedItem implements Special
         super(properties);
     }
 
-    abstract float getHeartsRequired();
+    abstract int getHeartsRequired();
 
     @Override
     public boolean handle(ServerPlayerEntity player, VillagerEntityMCA villager) {
